requirements.txt
torch
torchvision
onnx
onnxscript
onnxruntime
tqdm
scikit-learn
matplotlib
seaborn
pyyaml
pyzmq